Rename to neon-missing profile
